3.0.b
Added VLAN.active and VLAN.activate() to make 1 VLAN active at a time.
Dumps a shell script  in KSROOT that netadm or root needs to run.
Moved dev notes to README.md.
Code cleanup and reduction in vlan views.
This commit includes a model change, be sure to makemigrations/migrate.